refactor: move get_updated_feature_states to service layer and add multivariate support
- Moved get_updated_feature_states from EnvironmentFeatureVersion model to versioning_service
- Renamed to get_updated_feature_states_for_version for clarity
- Added multivariate percentage allocation change detection
- Moved 6 tests from model tests to service tests
- Added 2 new tests for multivariate changes (weight and value)
